WIP creating the login screen
